No need for whole static library
The linker prunes any unneeded symbols when you use
LOCAL_STATIC_LIBRARIES, but when you use LOCAL_WHOLE_STATIC_LIBRARIES it
tells the linker you don't want any pruning of the .a files to happen.
Since we only need what org_conscrypt_NativeCrypto.cpp calls, we can
use LOCAL_STATIC_LIBRARIES to allow the pruning to happen.

The order of the libraries in LOCAL_STATIC_LIBRARIES matters because if
a library earlier in the list has unresolved symbols, then the libraries
later in the list are checked to see if they have any of the unresolved
symbols.

Test: find out/ -name '*.apk' -exec rm {} \;; make -j32
Change-Id: I2fb27c5915d945a51bdf6c42f69174052f5bbe3c