NativeCrypto: add debugging format checking when debug off
Before if you enabled WITH_JNI_TRACE you might get some formatting
errors because format is not checked when debugging is enabled. Switch
to constexpr to enable debugging and rely on Dead Code Elimination pass
in the compiler to remove all the debug code when it's not in use. This
allows the compiler to properly check printf-style formatting for debug
statements instead of the preprocessor removing the code.

Test: compile with kWithJniTrace = true and run vogar tests
Change-Id: Ief3fe1c099a38d802db32deb7ffa91e4c8d4a572