Restrict TLS signing to non-RSA-PSS algorithms
A recent change in BoringSSL allowed connections to use RSA-PSS as the
signing algorithm for TLSv1.2 connections. However, the CryptoUpcalls
interface is not ready for this and it cannot currently make the
upcall correctly to have these signed.

Temporarily disable RSA-PSS signatures with TLS by explicitly setting
the list of signature algorithms.

Test: cts-tradefed run cts -m CtsLibcoreTestCases -a arm64-v8a
Test: cts-tradefed run cts -m CtsLibcoreOkHttpTestCases -a arm64-v8a
Bug: 31714503
Change-Id: Ie1c23b7231b5673816946a6c06030e1e25752415