Fix-up sslSelect logic for Windows
This allows Windows to work with parity for existing code. Interrupting
works and all the tests pass.

In the future, the #ifdefs could be split out into a different file
so we could include more strategies such as epoll or kqueue in the
future. However, this is only selecting on two objects.

Change-Id: Ia5c58b72e78aeac129f2987b99ce2a6d75448d32