Fix-up sslSelect logic for Windows (#41)
Fix-up sslSelect logic for Windows

This allows Windows to work with parity for existing code. Interrupting
works and all the tests pass.

In the future, the #ifdefs could be split out into a different file
so we could include more strategies such as epoll or kqueue in the
future. However, this is only selecting on two objects.