Specify the library path explicitly
Using "-lssl -lcrypto" can induce the linker to choose the
system-installed libraries instead of the compiled BoringSSL libraries
which will cause calamity when trying to run the resulting binaries.
Instead specify the library by the path explicitly which ensures that
the BoringSSL build intended is the one that is used.