Don't call SSL_get_state.
SSL_state_string_long provides a string which is more readable for
humans. We plan to unexport the numerical value to prevent code from
depending on implementation details of the state machine. The
string-based APIs will probably stay for debugging as they're much less
likely to be acted on programmatically.