Reduce verbosity of CI outputs (#79)
* Reduce verbosity of CI outputs

The build log is growing above the limits allowed for Travis-CI (10000
lines) with verbose output enabled. Since the build is fairly stable,
reduce the verbosity of the output for now.

This will allow us to see warnings a bit more clearly as well.

* Unify the way in which Android SDK is downloaded

Use sdkmanager in both Windows and Linux. This should reduce build
verbosity down a bit.