Share JVM buffers with native SSL functions.
The NativeCrypto SSL_read and SSL_write functions currently call
GetByteArrayElements and ReleaseByteArrayElements. According to the JNI
spec, This may result in a copy when data is transferred from the JVM to
native code and an additional copy if changes to the data need to be
written back to the JVM. On all machines I tested, this results in
copies (in both directions, when changes need to be committed).

This change uses GetPrimitiveArrayCritical and
ReleasePrimitiveArrayCritical, eliminating copies even when the JVM
implementation does not support pinning. It also introduces
ScopedCriticalByteArrayRO and ScopedCriticalByteArrayRW helper classes
which support memory access and release via RAII. They should only be
used within critical sections where no other JNI functions are called
and no syscalls block on the completion of other JNI calls. See
http://docs.oracle.com/javase/8/docs/technotes/guides/jni/spec/functions.html#GetPrimitiveArrayCritical_ReleasePrimitiveArrayCritical
for further details.