Minimize buffer copy penalty in SSL read/write. (#93)
* Minimize buffer copy penalty in SSL read/write.

This change is an attempt to address performance issues in
NativeCrypto.SSL_read and SSL_write when large buffer sizes are used.
It attempts to guess whether Get/ReleaseByteArrayElements will result in
copies between native code and the JVM. If so, Get/SetByteArrayRegion is
used instead, operating on small chunks of the target buffer and
ensuring that at most one copy per region crosses the JNI boundary.

This uses the same strategy as the cipher streaming code here:
https://github.com/google/conscrypt/blob/196ca9aad6223e2962e1330cff9e885c569c20ca/common/src/jni/main/cpp/NativeCrypto.cpp#L2410.

See https://github.com/google/conscrypt/pull/89 for additional context.

* Clear exception if bytes are available.

If an FD-closed exception is encountered after bytes have been
successfully read, clear the exception and attempt to return the partial
result. Rely on subsequent calls to sslRead regenerating the exception.