Make SSLUtils public for testing but @Internal
On Android the class-under-test is going to be in the bootclasspath
ClassLoader, but the test-class will be in the "app" ClassLoader. This
means that you can't directly test package-scoped classes or methods.

Make this public so we can test it on Android, but marked @Internal so
it doesn't end up in the Javadocs.

Test: cts-tradefed run cts -m CtsLibcoreTestCases -a arm64-v8a
Change-Id: I127693e1092fa9016c4a5d89218cc8de0c3421a5