Locking down public APIs
Tried to be as aggressive as I could, so this probably deserves a fairly thorough review.  I left most of OpenSSLSocketImpl public, because I think it's needed by a few external projects.

I also did some cleanup work to get rid of a bunch of compiler warnings that we seem to have accumulated.

Fixes #142