Don't fill in RSA_METHOD.encrypt when not needed.
This resolves a TODO, at the cost of adding two more. Newer BoringSSL
revisions ignore the hook as of
https://boringssl-review.googlesource.com/c/15864/.

It's probably unnecessary in older revisions too, but since Conscrypt
does not fill in rsa->n or rsa->e, it's plausible that, prior to that
BoringSSL change, someone might have passed Conscrypt's custom RSA
objects into a codepath that ultimately used RSA_encrypt. Tracing
through Java cryptography goo is probably not worth the trouble, so I've
left it as an #ifdef which can be removed in a couple weeks.