Move closer to the old test version
The original test never called shutdownInput() since the Javadoc
indicates it will make the socket behave in a different manner than
expected in this test.

Also assert that we're not using the type of socket that we expect to
return -1 if we get a SocketException. This will indicate when this
difference is fixed and the test can be changed to reflect the
expectation.