Use buffer allocator in engine
Also modifying the engine so that when running with no allocator
(i.e. default), that it lazy-creates a direct buffer for JNI operations.
This is a trade-off of performance for memory. It requires the
allocation of a single direct buffer (~16k) that it uses to copy
from/to heap buffers provided by the application.

Updated benchmarks are coming soon, but early results show a huge
improvement in both the engine and the engine-based socket.

Fixes #226