Pass encoded local certs to BoringSSL
Current code was encoding and then decoding the certs before finally
passing them to the native code. We were also separately setting
and then verifying the private key.  All of this can be replaced
with a single JNI call SSL_set_chain_and_key, which accepts the
encoded certs (we don't have to decode them again).

See #247

This shows a perf bump for the handshake (from ~750 to 800 ops/sec).