Fetch the peer certificates if necessary when handshake completes.
I'm not sure why, but under some circumstances the first connection on
a session can not have the cert_verify_callback called, which leaves us
without the peer certificates, causing hostname verification to fail in
apps (since they don't have a certificate to verify the hostname with).
This fixes the problem for the short term.