Encode certs in verify callback (#248)
This code was copied from Netty/netty-tcnative and seems to
significantly increases performance of the verify callback. Before
we call back to Java, we first encode all of the certs and then
decode them in Java into X509Certificate instances. Previous code
was calling into JNI for each method in the certificate.

This helps in addressing #247.