Switch to SSL_get0_peer_certificates.
This works towards issue #258. So the exception can be routed out
properly, this moves the SSL_get0_peer_certificates call to after
doHandshake completes in ConscryptFileDescriptorSocket.

Although, due to False Start (the "cut-through" logic in that class),
the handshake may not be fully complete at the time, BoringSSL's API is
such that the certificates and other properties will be available once
SSL_do_handshake first completes.