Implement Caliper benchmarks for running on Android devices (#10). (#278)
Netty tcnative doesn't work on ARM, so we split out the Netty-dependent
benchmark portions into benchmark-openjdk and put non-Netty versions
in benchmark-android.

The gradle operations required to execute the benchmarks on devices are ugly,
but they appear to work.  This might get better some time in the future
with improved Caliper support for Android.

This only includes the time-based benchmarks, not the throughput-based ones,
since those will require some more modifications to work in Caliper and these
at least get us started.