Allow beginHandshake() when a handshake is in process. (#293)
The documentation for beginHandshake() says that IllegalStateException is
thrown if the client/server mode hasn't been set, and it's reported that
the OpenJDK implementation and previous Android releases both allowed
calling beginHandshake() when a handshake is in progress.  Since
beginHandshake() is documented as being non-blocking, this should be safe.

This fixes https://issuetracker.google.com/37078438.