Support Java 6 Runtime
Various fixes to support Java 6, 7, and 8. Separating out
utility classes (for openjdk) to be explicit as to which methods are
supported by particular Java version.

Adding the ability to specify the test JVM to use on the command-line.

For example, the following will build with the default Java
installation, but will run the openjdk and integ-tests with Java 6:

./gradlew build -DjavaExecutable64=${JAVA6_HOME}/bin/java

Fixes #298