Account for OpenSSL 1.1.0 const-correcting a callback. (#296)
BoringSSL plans to switch to advertising OpenSSL 1.1.0's version number
at some point in the future. (We've been opaquifying the SSL stack which
leaves us in this awkward mix of 1.1.0 and 1.0.2 right now.) In one
case, it is unfortunately a breaking change because OpenSSL changed a
callback's type signature while keeping the name.

To account for that, add an OPENSSL_VERSION_NUMBER check. This makes
Conscrypt compatible with both BoringSSL today and a later BoringSSL. It
can be removed once the migration is completed. I'll send a PR to remove
this when that's happened.