Remove all remaining uses of SSL_clear.
I couldn't find anywhere that we're relying on the functionality behind
SSL_clear (indeed, the documentation of its behavior isn't very detailed
so I don't really know how we could), and BoringSSL recommends not using
it.  We allocate a separate SSL object per SSLEngine/SSLSocket, so we
aren't attempting to reuse them.  So this should be safe.