Switch R_RSA_DATA_TOO_LARGE{,_FOR_MODULUS} exception (#328)
These error codes can be generated by the methods used in both raw RSA
signatures and RSA ciphers, but the Java API for Cipher and Signature
has disjoint declared exceptions.  Luckily, in our Signature
implementations we catch any Exception and wrap it in a SignatureException,
so we can safely turn these into the appropriate exceptions for Cipher
and they'll be thrown as SignatureException from Signature anyway.