Suppress duplicative progress from sdkmanager
The 26.1.1 release of sdkmanager produces constant progress messages,
many of which are duplicates, which fills up the Travis logs
unnecessarily.  Replace the CRs (which make it look nice in the terminal)
with linefeeds, then use uniq to suppress any duplicated lines, which
means any given install will produce at most 100 lines in the
log instead of many thousands.