Throw SocketException if the FD socket has had its SSL freed (#343)
Under some circumstances (which are repeatable, but I haven't been able
to lock down), the HTTP connection reuse system in the JDK will attempt to
reuse a finalized ConscryptFileDescriptorSocket.  When this happens, the
socket throws a NullPointerException from the native code when methods
are called, because the native SSL object has been freed, which causes
a lot of problems.  Instead, have the socket throw SocketException instead,
which everything understands and responds properly to.

As best as I can tell, this is happening because the finalizer of some
object that's not ours but has a strong reference to our socket adds
a new strong reference, but our finalizer has already been enqueued
and thus is run even though the object then can later be reused.  I haven't
been able to find this finalizer, but everything tolerates the code as
written and responds properly to the SocketException, so it seems like a
good solution until we get rid of the FD socket entirely.

Fixes #331.