setSocketWriteTimeout(): Log more details in case of failure. (#345)
In response to a previous hard-to-diagnose bug, this method logged
a couple of stacktaces from the caught exception; but it didn't
log the exception type (e.g. InvocationTargetException), exception
message or causes. This CL expands this to also log the cause chain,
providing the exception type, message, and the first two
StackTraceElements for each Throwable in the chain.