Don't depend on openjdk classes in platform build.
The platform gradle build currently has a build dependency on openjdk
and doesn't include its own Platform class, so this works in the GitHub
build, but it doesn't work when it gets imported into Android's build
tree.  I'm working on a change to fix the build setup as well, but
platform's Platform.java depends on a lot of non-public functionality
in the Android world, so it's going to be more complicated than this.