Implement ChaCha20 support.
Adds Cipher, KeyGenerator, and AlgorithmParameters for ChaCha20.

Main changes are in OpenSSLCipher$EVP_AEAD.  Refactored so the GCM-
specific portions are all in the GCM subclass and the generic AEAD
portions (such as not allowing reuse of key/IV combinations) are in
the superclass.

Also updates so that calling Cipher.init() on an instance of
AES/GCM/NoPadding with a GCM AlgorithmParameters will pick up the
values from the AlgorithmParameters (previously, it would only work
with GCMParameterSpec or something convertable to IvParameterSpec).