Switch from wrapping to subclassing for Java 9 methods.
Older Android apps rely on OpenSSLSocketImpl being in the superclass
chain of sockets, which means wrapping them can cause problems.  Instead,
use a subclass to add the appropriate methods.  This also makes the
subclass implementation quite clean, but it pollutes the Platform classes
with a bunch of methods and means we have to remember not to use the
constructors directly.