Specify all TLS versions 1.0-1.2.
If you specify a set of TLS versions with a gap in them, all versions
above the gap are ignored, because the TLS protocol actually only
includes a single number (the highest protocol version that's allowable).
So the current version was specifying TLS 1.0 only.  This fixes it to use
whatever version is the highest available.