Implement unadorned ChaCha20.
ChaCha20 is a stream cipher in its own right, and it was pointed
out that it was weird that requesting "ChaCha20" alone from Conscrypt
returned an implementation of ChaCha20+Poly1305.  This implements
plain ChaCha20 and makes it the default implementation, so to access
ChaCha20+Poly1305 the caller must ask for ChaCha20/Poly1305/NoPadding
explicitly.

We haven't made a release with ChaCha20 in it yet, so it should be fine
to change the meaning of requesting "ChaCha20".