Check an X509-like structure submember for nullness.
We've seen very sporadic crashes due to null pointer dereferencing
somewhere inside X509_get_ext_by_critical, and this is the only way I
can see that that can happen.  X509_get_ext_by_critical passes
x->cert_info->extensions to X509v3_get_ext_by_critical, and that's the
only pointer that isn't explicitly checked for nullness.  These
crashes are incredibly rare, so it's not out of the realm of
possibility for them to be memory corruption or something, but better
safe than sorry.