Cache session values when SSL is being closed.
There's no documentation on what the behavior of an SSLSession object
should be after the connection it's associated with has been closed,
but callers reasonably expect that it remains callable, especially
since the connection can be closed by the remote peer without notice.
This makes the ActiveSession cache any values it extracts from the SSL
object locally before the object is freed, so that the getters can
continue to be called.  Prior to this, calling any getter whose value
wasn't already cached would result in a NullPointerException.