Update testing infrastructure.
This is part one of migrating the source of truth for tests from
Conscrypt to AOSP.  This sets up the basic layout we want to use:
tests under org.conscrypt.** that test the installed provider (or all
installed providers, for future crypto-related tests), and a separate
test suite class used by the Gradle build that installs the provider.
That way these tests can easily be used in any environment that can
properly install the provider to be tested, including AOSP.

This also updates the test classes with a few changes that had been
made to AOSP, mostly additional test cases.