Add crypto tests to Conscrypt.
These tests are pulled from AOSP with minor modifications.  They test
the behavior of the Cipher, KeyAgreement, and KeyGenerator classes for
all installed providers (which includes Conscrypt when run under the
ConscryptSuite).

Specific modifications made:

* Convert tests to JUnit4 style.
* Fixes for running under OpenJDK and with SunJCE installed.
* Removed tests for classes themselves rather than providers.

Note that these tests are not all up to modern standards in terms of
code style and readability, but I'd like to commit them as-is on the
basis that we can gradually improve them, but they'll do a lot more
work being available and accessible than they would stuck in AOSP
until they can be cleaned up.