Pass NativeSsl references to NativeCrypto (#408)
* Pass NativeSsl references to NativeCrypto

The existing implementation of passing raw addresses to NativeCrypto
can cause issues where the native code may still be executing when the
finalizer runs and frees the underlying native resources.  A call to
NativeSsl.read(), for instance, is not enough to keep the NativeSsl or
its owning socket alive, so if it's waiting for input the finalizer
can run.  Switching to passing the Java object to native code keeps
the Java object alive for GC purposes, preventing its finalizer from
running.

As part of this, also move the freeing of NativeSsl instances into a
finalizer on NativeSsl instead of on the sockets.  The sockets can
still become garbage even if the NativeSsl is kept alive, so we only
want to free it when the NativeSsl itself is garbage.

We will also want to do this for other native objects, but SSL*
instances are by far the most-used native objects and the most likely
to be used in a long-running I/O operation, so starting here gives us
a lot of benefit.

* Reliably close objects in tests.

* Pass both pointer and Java reference.

This allows us to access the SSL* pointer without having to indirect
through the Java object's fields, but still prevents the NativeSsl
from being GCed while the method is being run.

* Explain unsafe finalization fix in NativeCrypto Javadoc.