Always clear BoringSSL error when throwing exceptions.
We need to ensure that the BoringSSL error stack is clear when
returning from native functions, otherwise a later function might
inspect the error stack and interpret it incorrectly.  Since we always
return after throwing an exception, clear the error stack immediately
before throwing the exception.  This also fixes a number of places
where we failed to manually clear the error stack before returning.