Add compile option for checking error queue. (#416)
We need to ensure that the BoringSSL error stack is clear when
returning from native functions, otherwise a later function might
inspect the error stack and interpret it incorrectly.

Adds the compile option CONSCRYPT_CHECK_ERROR_QUEUE which enables the
macro CHECK_ERROR_QUEUE_ON_RETURN.  That macro, when enabled, creates
a class that checks the error queue is empty in its destructor.  The
macro has been added to almost every native method called from Java,
enforcing that the error queue is empty when we return from native
code back to Java code.

Adds the gradle property checkErrorQueue to enable the code and adds
it to the Travis config.

Also fixes a couple places found by this checking that we were failing
to clear the error queue after handling errors.