Add PSS AlgorithmParameters.
Upcoming tests for Signature require that PSS AlgorithmParameters are
available, and Conscrypt doesn't have an implementation.

Also refactors OAEPParameters so that the reading and writing of the
message hash and MGF1 hash are in their own functions, allowing the
code to be shared with PSSParameters.  The two are both specified in
RFC 4055 and are used for related purposes, so this is sensible from a
specification perspective in addition to from a DRY perspective.

One wrinkle in the PSSParameters implementation is that it only allows
reading and writing a value of 1 for the trailerField, whereas other
implementations allow any integer.  1 is the only legal value for this
field under RFC 4055 section 3.1 ("The value MUST be 1"), so any
compliant usage should be fine with this, and I think it's better to
refuse to handle any other value than let bad values move through the
system and explode later or be silently ignored.