Add tests for classes in java.security.
These tests are pulled from AOSP with minor modifications.  They test
the behavior of the CertificateFactory, MessageDigest, and Signature
classes, as well as a bunch of varieties of
AlgorithmParameterGenerator, AlgorithmParameters, KeyFactory, and
KeyPairGenerator classes for all installed providers (which includes
Conscrypt when run under the ConscryptSuite).

As usual, I've adapted them to JUnit4, removed the tests for the
classes themselves and the provider framework, and made them work
under OpenJDK as well as Android.

The quality of these tests is again variable, but that can be improved
once they're in and running.