Revert "Throw SocketException on ERR_SSL_SYSCALL. (#430)"
We've had reports from people that depended on "normal" connection
termination (such as in HTTP where the peer sends data and then closes
the TCP connection) returning -1 from read(), and in some cases this
change causes read() to throw a SocketException in that circumstance.

Rather than causing such a fundamental change in behavior, probably
better to just adjust the test to allow for a -1 return.

This reverts commit d84f6f57b216ab53dc3a71d30fd2252ce2208c13.