Calculate output size ourselves for AES/GCM and ChaCha20. (#438)
Since both AES/GCM and ChaCha20 don't have any variance in the length
of their AEAD tags, we can just calculate the output length directly
rather than making a JNI call.  This gives better bounds for the
necessary buffer size; in particular, we previously weren't factoring
in the removal of the tag during decryption, so we always demanded an
unnecessarily-long buffer.