Add -std=c++11 to constants build.
Some combinations of clang binary plus gcc headers result in errors on
the __float128 type, which is present in the gcc headers but
unsupported by clang.  Setting the -std flag makes clang set
preprocessor definitions such that the __float128 type doesn't show up
in the build.  We should be setting the desired standard library
anyway rather than leaving it up to the defaults.