Update short buffer handling. (#440)
This fixes a number of problems in Conscrypt's ciphers when they are
given an output buffer that is too small for the output.

We didn't specifically handle CIPHER_R_BUFFER_TOO_SMALL errors from
BoringSSL, so we threw RuntimeException on encountering them.  We
should be throwing ShortBufferException.

Raw ChaCha20 didn't check for a short buffer at all, so it just passed
the arrays down to native code, which could cause crashes or other
weird behavior.

EVP_AEAD ciphers used update() to only record data that needed to be
encrypted and then did the actual encrypting in doFinal().  This
combined with our implementation that implements doFinal() as a
combination of updateInternal() + doFinalInternal() led to a situation
where if the buffer passed to doFinal() was too small, the data would
get added to the buffer to be encrypted in updateInternal() and then
doFinalInternal() call would fail, which would mean a future call to
doFinal() with the same data would end up encrypting that data twice.
This call pattern is used by some internal CipherSpi methods from
OpenJDK, so you could see it in practice even if the caller did
nothing wrong.

Also add tests around short buffer handling.