Parse ASN1_TIME structures in constructors.
The legacy OpenSSL APIs in BoringSSL don't parse ASN1_TIME values
until they're used, which means that the existing code could explode
in the middle of X509Certificate.getNotAfter() and similar
Date-returning calls, and those calls aren't declared to throw
anything.  Instead, read and cache the values in the constructor,
where we can throw a relevant exception if necessary.

We have to clone the Date values when returning them because Date is
mutable.