Check for zero-length output buffers.
SSL_read()'s return value is the number of output bytes on success or
<= 0 on failure, with the failure details interpreted by calling
SSL_get_error().  When an empty output buffer is used, then a
successful operation can return 0, since there isn't room for any
output to be produced, but SSL_get_error() will still treat it like
the failure mode that a 0 result would indicate (specifically, an
unexpected EOF).  Check for this case and return a successful result.

Fixes #452.