Add logging macros that work on all platforms. (#462)
This adds CONSCRYPT_LOG_X macros that redirect to either ALOG on
Android or fprintf(stderr) on non-Android.  In the future, we could
use these to allow users to register a logging callback and send the
logs to a destination of their choice (via java.util.Logger or log4j
or what have you), but for now we'll keep it simple.

Fixes #460.