Remove CONSCRYPT_LOG_DEBUG.
We only had two places that CONSCRYPT_LOG_DEBUG was used, and they're
really errors, since they're failures in the code that throws Java
exceptions, which "shouldn't" happen.  In addition, _VERBOSE logging
can be suppressed, but _DEBUG can't, which isn't typical for those
names.  Just go with _VERBOSE logging if we need to log something in a
suppressable manner.