Disallow invalid SNI hostnames in setHostname(). (#470)
The code that sets the SNI value on the connection checks for

impl.getUseSni() && AddressUtils.isValidSniHostname(hostname)

which is good for hostnames that were supplied as the hostname to
connect to, since they may or may not be valid for SNI, but means that
if you set an invalid hostname with setHostname() it will just
silently be omitted from the connection and no SNI extension will be
included in the handshake.  Better to reject the hostname immediately.

Also disallow hostnames with trailing dots, which aren't legal SNI
hostnames per RFC 6066.

Also disallow null bytes.