Swap if conditions to avoid warning.
JDK9+ produces a warning when accessing the internal field that
hasFileDescriptor() checks to see if a file descriptor is available.
Users using the engine socket don't need to do that check, so swap the
order of if conditions so it comes second and gets avoided if the
engine socket is enabled.

Fixes #472.