Remove exception throwing on bad hostnames. (#478)
Testing on Android has indicated that there are too many things that
rely on passing whatever string has been passed to them to
setHostname() and aren't prepared to get an exception (including IP
addresses, single-element hostnames, and whatnot) for it to be
reasonable to throw an exception here.  Settle for documenting what
the behavior of passing an invalid hostname is.