Updates for new java.security tests. (#483)
Synchronize a number of Signature methods.  If these methods are
called concurrently, it can lead to crashes due to the native memory
management stepping on itself.

Add @RunWith(JUnit4.class) to the classes and @Test to a couple
missing places.

Fix the concurrent usage Signature test.  Previously it didn't init
the signature object, so all of the threads just immediately threw
without actually doing anything.  Also check the created Futures to
satisfy ErrorProne.