Add a function to force an engine read. (#453)
The SSLEngine implementation sometimes need to prompt BoringSSL
to process any incoming TLS data in order to determine whether
there is enough data to produce plaintext.  We previously were
reading into a zero-byte buffer to do this, but that causes
ambiguity in return values because a return of 0 from SSL_read()
could mean a failure (generally EOF) or could mean a successful
read of 0 bytes (which would be expected when reading into a
zero-byte array).  Instead, use SSL_peek.

Fixes #452.