Fix NPE from closed SSLEngine. (#499)
When calling wrap() on a closed SSLEngine instance, we check to see if
anything is already pending to be written, but the native BIO will
already have been freed, so the native code threw
NullPointerException.  Check for it being already freed when looking
for pending data and just return zero in that case.

Also sets the bio pointer to 0 before freeing it and and marks it as
volatile to reduce the chances of any weird race conditions, though
the locking in ConscryptEngine should already make these impossible.

Fixes netty/netty#7988.