Move TrustManagerImpl to common.
This is the first step in making the TrustManager implementation
available on OpenJDK, which is necessary for TLS 1.3 because the
TrustManager implementations in OpenJDK throw an exception when they
encounter an SSLSocket or SSLEngine that reports a protocol not in
their enumerated lists.  It moves the code and tests over so they
compile on OpenJDK, but doesn't yet change anything about the
provider, so the code isn't used.

There are a few changes necessary:

Extract an interface for TrustedCertificateStore, which is still
platform-only due to its reliance on specific directories to find
certs and other details of the Android platform.

Extract an interface for CertBlacklist, for the same reason.

Adjust TrustManagerImpl to get implementations of the add-on
functionality (CertBlacklist, CTLogStore, and CTPolicy) from the
Platform class instead of constructing them itself and tolerate nulls
for those classes.  The Platform classes for non-platform builds
return null for those classes, but we might want to change that in the
future.

Make a few minor changes while moving things over (JDK 6 source
compatibility, adjust tests for JUnit4, etc.).